refactor: add-room & manage-students depend on core-rooms (drop feature->feature dep)
