+++ shared/feature-add-room/impl/src/commonMain/kotlin/dev/nonoxy/residetrack/feature/add_room/impl/domain/AddRoomExecutor.kt
@@ -9,8 +9,8 @@ import dev.nonoxy.residetrack.feature.add_room.api.store.AddRoomSuccessKind
 import dev.nonoxy.residetrack.feature.add_room.impl.domain.AddRoomStoreFactory.Action
 import dev.nonoxy.residetrack.feature.add_room.impl.domain.AddRoomStoreFactory.AddRoomErrorKindOrNull
 import dev.nonoxy.residetrack.feature.add_room.impl.domain.AddRoomStoreFactory.Message
-import dev.nonoxy.residetrack.feature.rooms.api.models.Room
-import dev.nonoxy.residetrack.feature.rooms.api.repository.RoomsRepository
+import dev.nonoxy.residetrack.core.rooms.models.Room
+import dev.nonoxy.residetrack.core.rooms.repository.RoomsRepository
 import dev.nonoxy.residetrack.core.mvikotlin.BaseExecutor
 import kotlinx.collections.immutable.toPersistentList
 import kotlinx.coroutines.CoroutineDispatcher
