+++ shared/feature-manage-students/api/src/commonMain/kotlin/dev/nonoxy/residetrack/feature/manage_students/api/store/ManageStudentsStore.kt
@@ -4,7 +4,7 @@ import com.arkivanov.mvikotlin.core.store.Store
 import dev.nonoxy.residetrack.feature.manage_students.api.store.ManageStudentsStore.Intent
 import dev.nonoxy.residetrack.feature.manage_students.api.store.ManageStudentsStore.Label
 import dev.nonoxy.residetrack.feature.manage_students.api.store.ManageStudentsStore.State
-import dev.nonoxy.residetrack.feature.rooms.api.models.Room
+import dev.nonoxy.residetrack.core.rooms.models.Room
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf
